Change the recipe model to be more flexible.
Before each step had much information about the items, and it was not developer friendly this way. Therefore, every step only has an id or the name of the ingredient/tool instead of the full item object.

Furthermore, the recipe editor is now wayyyyy easier to use. Simply add new items and voilà! No need to go through all steps and manually add the items because it's not a requirement anymore to do so. The items are now in the same hierarchy as the steps.

This commit is def wip but I want a backup